pyproject.toml
[project]
name = "autism-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.116.2",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "uvicorn>=0.35.0",
]
